requirements.txt
netaddr
python-libnmap
pynetbox
pyyaml
requests
netmiko
